feat: реазиловать основную логику приложения

+++ ewm-server/src/main/java/ru/practicum/category/service/CategoryServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.practicum.category.service;
 
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -13,19 +14,24 @@ import ru.practicum.category.exception.CategoryException;
 import ru.practicum.category.mapper.CategoryMapper;
 import ru.practicum.category.model.Category;
 import ru.practicum.category.repository.CategoryRepository;
+import ru.practicum.event.repository.EventRepository;
+import ru.practicum.exception.ConflictException;
 
 import java.util.List;
 
 @Service
 @AllArgsConstructor
 @Transactional(readOnly = true)
+@Slf4j
 public class CategoryServiceImpl implements CategoryService {
     public static final Sort SORT_BY_ASC = Sort.by(Sort.Direction.ASC, "id");
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     @Transactional
     public CategoryDto add(CategoryDtoIn categoryDtoIn) {
+        log.info("Добавление новой категории {}", categoryDtoIn);
         Category category = CategoryMapper.makeCategory(categoryDtoIn);
         return CategoryMapper.makeCategoryDto(categoryRepository.save(category));
     }
@@ -33,6 +39,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public CategoryDto update(Long catId, CategoryDtoIn categoryDtoIn) {
+        log.info("Update category {}", categoryDtoIn);
         Category newCategory = categoryRepository.findById(catId).orElseThrow(() ->
                 new CategoryException("Category with id=" + catId + " was not found"));
         newCategory.setName(categoryDtoIn.getName());
@@ -42,14 +49,19 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public void delete(Long catId) {
-        //todo
-        Category deleteCategory = categoryRepository.findById(catId).orElseThrow(() ->
+        log.info("Delete category with ID {}", catId);
+        categoryRepository.findById(catId).orElseThrow(() ->
                 new CategoryException("Category with id=" + catId + " was not found"));
+
+        if (!eventRepository.findAllByCategoryId(catId).isEmpty()) {
+            throw new ConflictException("К категории привязаны события");
+        }
         categoryRepository.deleteById(catId);
     }
 
     @Override
     public List<CategoryDto> getAll(int from, int size) {
+        log.info("Получения всех ктегорий постранично {} и {}", from, size);
         Pageable pageable = PageRequest.of(from/size, size, SORT_BY_ASC);
         Page<Category> page = categoryRepository.findAll(pageable);
         return CategoryMapper.makeCategoryDtoList(page);
@@ -57,6 +69,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDto findById(Long catId) {
+        log.info("Get category with ID {}", catId);
         Category category = categoryRepository.findById(catId).orElseThrow(() ->
                 new CategoryException("Category with id=" + catId + " was not found"));
         return CategoryMapper.makeCategoryDto(category);
